Fix server error on account delete
This was caused by the pre-delete hook that is supposed to delete
the next_notification Schedule instance. However, the schedule is
deleted automatically if the user deletes the account herself, the
hook was only necessary for deletion via Admin interface, which
isn't particularly important, so we simply deleted the hook.